cloud_functions/phase_06_mesh/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
open3d==0.18.*
numpy==1.*
